filmsearch/requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
djangorestframework==3.14.0
django-filter==23.5
WeasyPrint==56.1 